+++ ikarus/finiteelements/mechanics/materials.hh
@@ -9,9 +9,8 @@
 
 #pragma once
 
-#include <ikarus/finiteelements/mechanics/materials/hyperelasticity/hyperelasticity.hh>
+#include <ikarus/finiteelements/mechanics/materials/hyperelastic/factory.hh>
 #include <ikarus/finiteelements/mechanics/materials/linearelasticity.hh>
-#include <ikarus/finiteelements/mechanics/materials/neohooke.hh>
 #include <ikarus/finiteelements/mechanics/materials/svk.hh>
 #include <ikarus/finiteelements/mechanics/materials/tags.hh>
 #include <ikarus/finiteelements/mechanics/materials/vanishingstrain.hh>
