+++ ikarus/finiteelements/mechanics/materials/hyperelastic/CMakeLists.txt
@@ -1,0 +1,10 @@
+# SPDX-FileCopyrightText: 2021-2024 The Ikarus Developers mueller@ibb.uni-stuttgart.de
+# SPDX-License-Identifier: LGPL-3.0-or-later
+
+# install headers
+install(
+  FILES factory.hh concepts.hh interface.hh neohooke.hh
+  DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/ikarus/finiteelements/mechanics/materials/hyperelastic
+)
+add_subdirectory(deviatoric)
+add_subdirectory(volumetric)
