+++ ikarus/finiteelements/mechanics/materials/hyperelastic/concepts.hh
