+++ ikarus/finiteelements/mechanics/materials/hyperelastic/deviatoric/CMakeLists.txt
@@ -1,0 +1,9 @@
+# SPDX-FileCopyrightText: 2021-2024 The Ikarus Developers mueller@ibb.uni-stuttgart.de
+# SPDX-License-Identifier: LGPL-3.0-or-later
+
+# install headers
+install(
+  FILES interface.hh ogden.hh blatzko.hh
+  DESTINATION
+    ${CMAKE_INSTALL_INCLUDEDIR}/ikarus/finiteelements/mechanics/materials/hyperelastic/deviatoric
+)
