+++ ikarus/finiteelements/mechanics/materials/hyperelastic/deviatoric/ogden.hh
