+++ ikarus/finiteelements/mechanics/materials/hyperelastic/neohooke.hh
